requirements.txt
faust-streaming==0.9.5

uvicorn==0.23.2
fastapi==0.103.1
gunicorn==21.2.0
matplotlib==3.8.3

clickhouse-sqlalchemy==0.3.0
asynch==0.2.3